blog/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
category==0.1.0
Django==3.0
djangorestframework==3.13.1
numpy==1.23.3
Pygments==2.13.0
pytz==2022.2.1
rust_category==0.2.0
sqlparse==0.4.2
